// apps/dashboard/src/app/(clean-code)/(sales)/types.ts
export interface StepComponentMeta {
    stepSequence?: { id?: number }[];
    deleted?: { [uid in string]: boolean };
    show?: { [uid in string]: boolean };
    variations?: {
        rules: {
            stepUid: string;
            operator: "is" | "isNot";
            componentsUid: string[];
        }[];
    }[];
    sortIndex?;
    sectionOverride?: {
        hasSwing?: boolean;
        noHandle?: boolean;
        overrideMode?: boolean;
    };
}
// ...
export interface StepComponentForm {
    title: string;
    id?;
    img?: string;
    stepId;
    isDoor?: boolean;
    custom?: boolean;
    productCode?: string;
}